+++ test/server_test.rb
@@ -4,9 +4,10 @@ require 'cucumber'
 require 'test_helper'
 require 'webrick'
 require_relative '../lib/maze/servlets/base_servlet'
+require_relative '../lib/maze/servlets/servlet'
 require_relative '../lib/maze/servlets/command_servlet'
 require_relative '../lib/maze/servlets/log_servlet'
-require_relative '../lib/maze/servlets/servlet'
+require_relative '../lib/maze/servlets/trace_servlet'
 require_relative '../lib/maze/server'
 
 # noinspection RubyNilAnalysis
@@ -43,6 +44,7 @@ module Maze
       mock_http_server.expects(:mount).with('/builds', any_parameters).once
       mock_http_server.expects(:mount).with('/uploads', any_parameters).once
       mock_http_server.expects(:mount).with('/sourcemap', any_parameters).once
+      mock_http_server.expects(:mount).with('/traces', any_parameters).once
       mock_http_server.expects(:mount).with('/react-native-source-map', any_parameters).once
       mock_http_server.expects(:mount).with('/command', any_parameters).once
       mock_http_server.expects(:mount).with('/logs', any_parameters).once
@@ -92,6 +94,7 @@ module Maze
       mock_http_server.expects(:mount).with('/builds', any_parameters).once
       mock_http_server.expects(:mount).with('/uploads', any_parameters).once
       mock_http_server.expects(:mount).with('/sourcemap', any_parameters).once
+      mock_http_server.expects(:mount).with('/traces', any_parameters).once
       mock_http_server.expects(:mount).with('/react-native-source-map', any_parameters).once
       mock_http_server.expects(:mount).with('/logs', any_parameters).once
       mock_http_server.expects(:mount).with('/command', any_parameters).once